Refactor slides file for easier maintainence
