+++ src/manim_beamer/blocks.py
@@ -1,4 +1,4 @@
-from abc import abstractmethod, ABC
+from abc import ABC, abstractmethod
 from typing import Union
 
 from manim import (
